feat(call the gifs function)
This function selects the seach button and input of the form element, once the search button is clicked the random gif will be displayed making use of the input by the user, if no gif can be find, a dog gif will appear. A catch error message was added to inform the user of any possible error.